verification-context-domain/src/main/java/uk/co/mruoc/idv/lockout/domain/model/RecordOnSequenceComplete.java
package uk.co.mruoc.idv.lockout.domain.model;

import uk.co.mruoc.idv.lockout.domain.service.RecordAttemptRequest;
import uk.co.mruoc.idv.verificationcontext.domain.model.VerificationContext;
import uk.co.mruoc.idv.verificationcontext.domain.model.result.VerificationResult;

public class RecordOnSequenceComplete implements RecordAttemptStrategy {

    public static final String TYPE = "on-sequence-complete";

    @Override
    public String getType() {
        return TYPE;
    }

    @Override
    public boolean shouldRecordAttempt(final RecordAttemptRequest request) {
        final VerificationContext context = request.getContext();
        final VerificationResult result = request.getResult();
        return context.containsCompleteSequenceContainingMethod(result.getMethodName());
    }

}
